requirements.txt
datamule>=0.342
pyyaml>=6.0
requests>=2.31.0
python-dateutil>=2.8.2